Changed folders and a few variables and map completed in main
Interfaces required me to change from float to int so I did, also completed the map

+++ TDGameUnity/Assets/extraScripts/Interfaces.cs
@@ -9,5 +9,5 @@ public class Interfaces : MonoBehaviour
 
 public interface IDamageable
 {
-    void TakeDamage(float damage);
+    void TakeDamage(int damage);
 }
